chore(menu): minor file reorganization
Reorganizes the files under `material/menu` so they make a bit more sense. The file for the `mat-menu` **component** isn't called `menu-directive` anymore. Also replaces the unnecessary `menu.ts` which was only re-exporting symbols.
