+++ src/lib/menu/public-api.ts
@@ -6,8 +6,11 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
+export {MatMenu, MatMenuDefaultOptions, MAT_MENU_DEFAULT_OPTIONS} from './menu';
+export {MatMenuItem} from './menu-item';
+export {MatMenuTrigger, MAT_MENU_SCROLL_STRATEGY} from './menu-trigger';
+export {MatMenuPanel} from './menu-panel';
 export * from './menu-module';
-export * from './menu';
 export * from './menu-animations';
 export * from './menu-content';
-export {MAT_MENU_SCROLL_STRATEGY} from './menu-trigger';
+export * from './menu-positions';
